refactor: rename: MemberUtils 이름 변경
VerifyMemberUtils 로 이름 변경

[skip ci]
