+++ server/edusync/src/main/java/com/codestates/edusync/model/study/postcomment/controller/StudygroupPostCommentController.java
@@ -1,12 +1,12 @@
 package com.codestates.edusync.model.study.postcomment.controller;
 
-import com.codestates.edusync.model.common.utils.MemberUtils;
+import com.codestates.edusync.model.common.utils.VerifyMemberUtils;
 import com.codestates.edusync.model.member.entity.Member;
 import com.codestates.edusync.model.study.postcomment.entity.StudygroupPostComment;
 import com.codestates.edusync.model.study.postcomment.mapper.StudygroupPostCommentMapper;
 import com.codestates.edusync.model.study.postcomment.service.StudygroupPostCommentService;
 import com.codestates.edusync.model.study.postcomment.dto.StudygroupPostCommentDto;
-import com.codestates.edusync.model.common.controller.UriCreator;
+import com.codestates.edusync.model.common.utils.UriCreator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -25,7 +25,7 @@ import java.util.List;
 public class StudygroupPostCommentController {
     private final StudygroupPostCommentService studygroupPostCommentService;
     private final StudygroupPostCommentMapper mapper;
-    private final MemberUtils memberUtils;
+    private final VerifyMemberUtils verifyMemberUtils;
 
     private static final String DEFAULT_STUDYGROUP_URL = "/studygroup";
     private static final String DEFAULT_STUDYGROUP_POST_COMMENT_URL = "/comment";
@@ -40,7 +40,7 @@ public class StudygroupPostCommentController {
     public ResponseEntity postStudygroupPostComment(@PathVariable("studygroup-id") @Positive Long studygroupId,
                                                     @Valid @RequestBody StudygroupPostCommentDto.Post postDto,
                                                     Authentication authentication) {
-        Member loginMember = memberUtils.getLoggedIn(authentication);
+        Member loginMember = verifyMemberUtils.getLoggedIn(authentication);
 
         StudygroupPostComment createdStudygroupPostComment =
                 studygroupPostCommentService.createStudygroupPostComment(
@@ -68,7 +68,7 @@ public class StudygroupPostCommentController {
                                                      @PathVariable("comment-id") @Positive Long commentId,
                                                      @Valid @RequestBody StudygroupPostCommentDto.Patch patchDto,
                                                      Authentication authentication) {
-        Member loginMember = memberUtils.getLoggedIn(authentication);
+        Member loginMember = verifyMemberUtils.getLoggedIn(authentication);
 
         StudygroupPostComment updatedStudygroupPostComment =
                 studygroupPostCommentService.updateStudygroupPostComment(
@@ -105,7 +105,7 @@ public class StudygroupPostCommentController {
     public ResponseEntity deleteStudygroupPostComment(@PathVariable("studygroup-id") @Positive Long studygroupId,
                                                       @PathVariable("comment-id") @Positive Long commentId,
                                                       Authentication authentication) {
-        Member loginMember = memberUtils.getLoggedIn(authentication);
+        Member loginMember = verifyMemberUtils.getLoggedIn(authentication);
 
         studygroupPostCommentService.deleteStudygroupPostComment(studygroupId, commentId, loginMember);
 
@@ -120,7 +120,7 @@ public class StudygroupPostCommentController {
     @DeleteMapping(DEFAULT_STUDYGROUP_URL + "/{studygroup-id}/all")
     public ResponseEntity deleteAllStudygroupPostComment(@PathVariable("studygroup-id") @Positive Long studygroupId,
                                                          Authentication authentication) {
-        Member loginMember = memberUtils.getLoggedIn(authentication);
+        Member loginMember = verifyMemberUtils.getLoggedIn(authentication);
 
         studygroupPostCommentService.deleteAllStudygroupPostCommentByStudygroupId(studygroupId, loginMember);
