+++ server/edusync/src/main/java/com/codestates/edusync/model/study/studygroup/controller/StudygroupController.java
@@ -1,8 +1,8 @@
 package com.codestates.edusync.model.study.studygroup.controller;
 
 import com.codestates.edusync.model.common.dto.MultiResponseDto;
-import com.codestates.edusync.model.common.utils.MemberUtils;
-import com.codestates.edusync.model.common.controller.UriCreator;
+import com.codestates.edusync.model.common.utils.VerifyMemberUtils;
+import com.codestates.edusync.model.common.utils.UriCreator;
 import com.codestates.edusync.model.member.entity.Member;
 import com.codestates.edusync.model.study.studygroup.entity.Studygroup;
 import com.codestates.edusync.model.study.studygroup.mapper.StudygroupMapper;
@@ -30,7 +30,7 @@ public class StudygroupController {
     private static final String STUDYGROUP_DEFAULT_URI = "/studygroup";
     private final StudygroupMapper studygroupMapper;
     private final StudygroupService studygroupService;
-    private final MemberUtils memberUtils;
+    private final VerifyMemberUtils verifyMemberUtils;
 
     /**
      * 스터디 모집 & 등록
@@ -41,7 +41,7 @@ public class StudygroupController {
     public ResponseEntity postStudygroup(Authentication authentication,
                                          @Valid @RequestBody StudygroupDto.Post postDto) {
 
-        Member member = memberUtils.getLoggedIn(authentication);
+        Member member = verifyMemberUtils.getLoggedIn(authentication);
         Studygroup studygroup = studygroupMapper.StudygroupDtoPostToStudygroup(postDto, member);
         studygroup = studygroupService.createStudygruop(studygroup);
         URI location = UriCreator.createUri(STUDYGROUP_DEFAULT_URI, studygroup.getId());
